xmlsec: handle missing EqualDistinguishedNames
I didn't try to do a GPG only build, but the standard non-NSS
build fails now because of missing a EqualDistinguishedNames,
which is non-trivial to implement.

This is the consequence of commit
1d3da3486d827dd5e7a3bf1c7a533f5aa9860e42 "xmlsecurity: add
EqualDistinguishedNames()" and commit
5af5ea893bcb8a8eb472ac11133da10e5a604e66 "xmlsecurity: improve
handling of multiple certificates per X509Data".

One could also use OpenSSL i2d_X509_NAME and X509_NAME_cmp to
program this function, but it looks like it requires some
ANS.1 input to begin with (maybe i2s_ASN1_* can be used).

Change-Id: I0600f76a5cb4379db428f67cb136b8567dfcc51e