gbuild: create full services.rdb from build
Currently the build handles global build dependencies redundant
in a few places:

* in Repository*mk + modules - the "real" build dependencies
* the full services.rdb generation
* the static UNO constructor map generation

As a result, the component files don't reflect the build components
and the RDB services generation redefines the build dependencies.

So this drops the latter two by generating the list of components
and it's constructors from the build itself. As one of the main
consequences, component files must now be split, so they reflect
the real components in the libraries.

This also adds a little convenience helper to add multiple
component files: gb_Library_set_componentfiles

Change-Id: I1b38a6f2c1e5221f18d7e5e756c30263b555d962